backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic-settings>=2.2
motor>=3.3
python-multipart>=0.0.9
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
httpx>=0.27.0
pdfminer.six>=20231228
python-docx>=1.1.2
antiword==0.1.0
